Move AbstractGuavaIterator into utils as in ASF
